Feature: Getting Network interface name and its RX/TX Speed
Feature: Getting Network interface name and its RX/TX Speed
- Get network interface list from /proc/net/dev
- Get RX/TX bytes and the number of drop / error packet of each interfaces from the file.
- Calculate RX/TX speed of each interfaces.
- If docker vritual interface (veth- / br-) exist, convert its ID to container / docker bridge Name.
- Code Refactoring of some function.

+++ info_to_log.c
@@ -185,8 +185,8 @@ void get_Temperature_Perccli(TempInfo* tempBuf){ // Get temperature information 
 }
 
 void get_CPU_Usage_Percent_of_All_Core(CpuUsage* cpuUsageBuf){ // Get CPU Usage(%) from Jiffies
-    unsigned long user = 0, nice = 0, system = 0, idle = 0, iowait = 0, irq = 0, softirq = 0, steal = 0, guest = 0, guest_nice = 0;
-    static unsigned long priv_total = 0, priv_idle = 0;
+    size_t user = 0, nice = 0, system = 0, idle = 0, iowait = 0, irq = 0, softirq = 0, steal = 0, guest = 0, guest_nice = 0;
+    static size_t priv_total = 0, priv_idle = 0;
     FILE *fp = NULL;
     cpuUsageBuf->usage = -100; // Initialization; -100 means "N/A"
